added controller for exercises
