+++ app/mongoBD/mongoConnection.ts
@@ -1,0 +1,7 @@
+import mongoose from "mongoose";
+
+const connectToMongo = async () => {
+  await mongoose.connect(process.env.MONGODB_URI!);
+};
+
+export default connectToMongo;
